APPS-5655 replacing recursion with iterations - fix compile errors

+++ common-jre/src/test/java/org/geogebra/common/euclidian/plot/interval/EuclidianViewBoundsMock.java
@@ -106,12 +106,12 @@ public class EuclidianViewBoundsMock implements EuclidianViewBounds {
 	}
 
 	@Override
-	public int toScreenCoordXd(double x) {
+	public double toScreenCoordXd(double x) {
 		return x;
 	}
 
 	@Override
-	public int toScreenCoordYd(double y) {
+	public double toScreenCoordYd(double y) {
 		return y;
 	}
